Added gradient analysis
- Better error handling
- Improved logic when wave propagates in the opposite direction

+++ src/graph_analyzer.py
@@ -1,7 +1,7 @@
 import csv
+from tkinter import filedialog, messagebox
 
 import numpy as np
-from tkinter import filedialog, messagebox
 from scipy.ndimage import gaussian_filter1d
 from scipy.signal import convolve2d
 
@@ -18,7 +18,7 @@ class GraphAnalyzer:
         """
         Compute the moving average of a 1D array using a specified window size.
         """
-        if self.mode == 2:
+        if self.mode == 3:
             return gaussian_filter1d(data, self.sigma)
 
         data = np.array(data)
@@ -31,9 +31,11 @@ class GraphAnalyzer:
             np.convolve(data, np.ones(self.window_size) / self.window_size,
                         mode='valid'))
 
-        if self.mode == 0: return res
-        if self.mode == 1: return res * - np.gradient(res)
+        if self.mode == 0: return -np.gradient(res)
+        if self.mode == 1: return res
+        if self.mode == 2: return res * -np.gradient(res)
 
+        # garbage
         return data
 
     @staticmethod
@@ -66,33 +68,37 @@ class GraphAnalyzer:
         Runs ``_max_sum`` twice, once in an inverted order to allow a strictly
         decreasing selection.
         """
+        is_increasing = (self._max_sum(l, 0.)[1] >
+                         self._max_sum([line[::-1] for line in l], 0.)[1])
         increasing = self._max_sum(l, weight_factor)
-        tmp = self._max_sum(l[::-1], weight_factor)
-        decreasing = tmp[0][::-1], tmp[1]
-        res = np.array(max(increasing, decreasing, key=lambda x: x[1])[0])
-        error = self.window_size // 2 if self.mode == 0 else 0
+        tmp = self._max_sum([line[::-1] for line in l], weight_factor)
+        decreasing = [(entry[0], len(l[0]) - entry[1]) for entry in tmp[0]], tmp[1]
+        res = increasing if is_increasing else decreasing
+        res = np.array(res[0])
+        error = self.window_size // 2 if self.mode != 3 else 0
         res = [(k[0] + 1, k[1] + error) for k in res]
         self.last = res
 
-        try:
-            file_path = filedialog.asksaveasfilename(
-                defaultextension=".csv",
-                filetypes=[("CSV files", "*.csv")])
+        if messagebox.askyesno("Save File", "Save output as a CSV file?"):
+            try:
+                file_path = filedialog.asksaveasfilename(
+                    defaultextension=".csv",
+                    filetypes=[("CSV files", "*.csv")])
 
-            if file_path:
-                with open(file_path, mode='w', newline='') as file:
-                    writer = csv.writer(file)
-                    writer.writerow(("Image number",
-                                     "Number of pixels from origin"))
-                    writer.writerows(res)
+                if file_path:
+                    with open(file_path, mode='w', newline='') as file:
+                        writer = csv.writer(file)
+                        writer.writerow(("Image number",
+                                         "Number of pixels from origin"))
+                        writer.writerows(res)
 
-                messagebox.showinfo("Success", "Image saved at"
-                                               f"\n{file_path}\n"
-                                               "successfully.")
+                    messagebox.showinfo("Success", "Image saved at"
+                                                   f"\n{file_path}\n"
+                                                   "successfully.")
 
-        except OSError as e:
-            messagebox.showerror(
-                "Error", f"An error occurred while saving graphs:\n{str(e)}")
+            except OSError as e:
+                messagebox.showerror(
+                    "Error", f"An error occurred while saving graphs:\n{str(e)}")
 
         return res
